+++ lib/eda/member.ex
@@ -14,6 +14,7 @@ defmodule EDA.Member do
     :deaf,
     :mute,
     :pending,
+    :permissions,
     :communication_disabled_until
   ]
 
@@ -29,6 +30,7 @@ defmodule EDA.Member do
           deaf: boolean() | nil,
           mute: boolean() | nil,
           pending: boolean() | nil,
+          permissions: String.t() | nil,
           communication_disabled_until: String.t() | nil
         }
 
@@ -46,6 +48,7 @@ defmodule EDA.Member do
       deaf: raw["deaf"],
       mute: raw["mute"],
       pending: raw["pending"],
+      permissions: raw["permissions"],
       communication_disabled_until: raw["communication_disabled_until"]
     }
   end